:recycle: Move updateTransactionsIfNeeded to transactions actions
