+++ jest.config.js
@@ -32,7 +32,7 @@ module.exports = {
     '^.+\\.css$': 'identity-obj-proxy',
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': '<rootDir>/__mocks__/fileMock.js',
   },
-  collectCoverage: true,
+  collectCoverage: !process.env.NO_COV,
   coverageDirectory: '<rootDir>/coverage/jest',
   collectCoverageFrom: process.env.NO_COV ? [] : [
     'src/**/*.js',
