+++ src/actions/account.js
@@ -1,12 +1,10 @@
-/* eslint-disable max-lines */
-// TODO figure out how to reduce size of this file
 import i18next from 'i18next';
 import actionTypes from '../constants/actions';
 import { getAccount, setSecondPassphrase } from '../utils/api/account';
-import { registerDelegate, getDelegate, getAllVotes } from '../utils/api/delegate';
+import { registerDelegate, getDelegate } from '../utils/api/delegate';
 import { getTransactions } from '../utils/api/transactions';
 import { getBlocks } from '../utils/api/blocks';
-import { transactionsUpdated } from './transactions';
+import { updateTransactionsIfNeeded } from './transactions';
 import { delegateRegisteredFailure } from './delegate';
 import { secondPassphraseRegisteredFailure } from './secondPassphrase';
 import { liskAPIClientUpdate } from './peers';
@@ -161,27 +159,6 @@ export const loadDelegate = ({ publicKey }) =>
         type: actionTypes.updateDelegate,
       });
     });
-  };
-
-export const updateTransactionsIfNeeded = ({ transactions, account }, windowFocus) =>
-  (dispatch) => {
-    const hasRecentTransactions = txs => (
-      txs.confirmed.filter(tx => tx.confirmations < 1000).length !== 0 ||
-      txs.pending.length !== 0
-    );
-
-    if (windowFocus || hasRecentTransactions(transactions)) {
-      const { filter, customFilters } = transactions;
-      const address = transactions.account ? transactions.account.address : account.address;
-
-      dispatch(transactionsUpdated({
-        pendingTransactions: transactions.pending,
-        address,
-        limit: 25,
-        filter,
-        customFilters,
-      }));
-    }
   };
 
 export const accountDataUpdated = ({
