+++ src/actions/account.test.js
@@ -1,5 +1,3 @@
-/* eslint-disable max-lines */
-// TODO figure out how to reduce size of this file
 import { expect as chaiExpect } from 'chai';
 import { spy, stub } from 'sinon';
 import i18next from 'i18next';
@@ -13,7 +11,6 @@ import {
   passphraseUsed,
   loadDelegate,
   accountDataUpdated,
-  updateTransactionsIfNeeded,
   updateDelegateAccount,
   delegateStatsLoaded,
   updateAccountDelegateStats,
@@ -302,44 +299,6 @@ describe('actions: account', () => {
 
       accountDataUpdated(data)(dispatch, getState);
       chaiExpect(peersActionsStub).to.have.been.calledWith({ online: false, code: 'EUNAVAILABLE' });
-    });
-  });
-
-  describe('updateTransactionsIfNeeded', () => {
-    let transactionsActionsStub;
-    let getState;
-
-    const dispatch = spy();
-
-    beforeEach(() => {
-      transactionsActionsStub = spy(transactionsActions, 'transactionsUpdated');
-      getState = () => ({
-        peers: { liskAPIClient: {} },
-      });
-    });
-
-    afterEach(() => {
-      transactionsActionsStub.restore();
-    });
-
-    it('should update transactions when window is in focus', () => {
-      const data = {
-        transactions: { confirmed: [{ confirmations: 10 }], pending: [] },
-        account: { address: accounts.genesis.address },
-      };
-
-      updateTransactionsIfNeeded(data, true)(dispatch, getState);
-      chaiExpect(transactionsActionsStub).to.have.been.calledWith();
-    });
-
-    it('should update transactions when there are no recent transactions', () => {
-      const data = {
-        transactions: { confirmed: [{ confirmations: 10000 }], pending: [{ id: '123' }] },
-        account: { address: accounts.genesis.address },
-      };
-
-      updateTransactionsIfNeeded(data, false)(dispatch, getState);
-      chaiExpect(transactionsActionsStub).to.have.been.calledWith();
     });
   });
